new script and functionality for text backing
