+++ .gitattributes
@@ -1,0 +1,1 @@
+*.mov filter=lfs diff=lfs merge=lfs -text
