+++ Assets/CohortUnityClient/Materials/Materials.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 7c387c41b7a1744a3b4a8177b0f84cd0
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
