+++ Assets/CohortUnityClient/Materials/Materials/video-rendertex.mat.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: e2e78810510c34b4cae454e3dc12fff7
+NativeFormatImporter:
+  externalObjects: {}
+  mainObjectFileID: 2100000
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
