+++ Assets/CohortUnityClient/Materials/Materials/video-rendertex1.mat.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 7c0bce9030ff948adafa63e6507aef51
+NativeFormatImporter:
+  externalObjects: {}
+  mainObjectFileID: 2100000
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
