+++ Assets/CohortUnityClient/Materials/video-rendertex.renderTexture.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 036396b6f8bd04a84989054bddaece03
+NativeFormatImporter:
+  externalObjects: {}
+  mainObjectFileID: 8400000
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
